requirements.txt
django
psycopg2-binary
djangorestframework
django-environ
django_filter
django-cors-headers